// src/front/src/api/identity/models/teamListFilter.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX Identity API (via Gateway)
 * OpenAPI spec version: 1.0
 */
import type { TeamType } from "./teamType";

export interface TeamListFilter {
	pageNumber?: number;
	pageSize?: number;
	readonly skip?: number;
	/** @nullable */
	search?: string | null;
	/** @nullable */
	structureId?: number | null;
	/** @nullable */
	isActive?: boolean | null;
	type?: TeamType;
}